refactor gradio UI
